Add files via upload
Functional Data Analysis (FDA) is a branch of statistics that focuses on analyzing and modeling data that are represented as curves or functions over a continuous domain, rather than discrete data points. It's useful for studying patterns in datasets where observations are curves, such as growth trajectories or time-series signals. Techniques in FDA include Functional Principal Component Analysis and Functional Regression among others.